rename: chatmessage를 chat-message로 통일
