+++ backend/src/models/chat-message.ts
@@ -1,4 +1,5 @@
 class ChatMessage {
+  id!: number;
   chatroomId!: string;
   userId!: string;
   content!: string;
