+++ backend/src/services/chat-message.service.ts
@@ -1,28 +1,28 @@
 import HTTPError from '../errors/http-error';
-import chatmessageQuery from '../query/chatmessage.query';
+import chatMessageQuery from '../query/chat-message.query';
 import { CreateChatmessageRequest } from "../requests/chatmessage.request"
 
-class ChatmessageService {
+class ChatMessageService {
 
   async create(creaetChatmessageRequest: CreateChatmessageRequest) {
     const { chatroomId, userId, content } = creaetChatmessageRequest;
 
-    const createChatmessage = await chatmessageQuery.create({
+    const createChatmessage = await chatMessageQuery.create({
       chatroomId,
       userId,
       content
     });
-    if (!createChatmessage ) {
+    if (!createChatmessage) {
       throw new HTTPError(500, "채팅룸 생성 실패")
     }
     return createChatmessage;
   }
 
   async findByChatrommId(chatrommId: string) {
-    const findChatrooms = await chatmessageQuery.findByChatroomId(chatrommId)
-    
+    const findChatrooms = await chatMessageQuery.findByChatroomId(chatrommId)
+
     return findChatrooms
   }
 }
 
-export default new ChatmessageService();
+export default new ChatMessageService();
